requirements.txt
fastapi
rank-bm25
langsmith
openai
langchain
openai-responses
python-dotenv